# [POC][DEPRECATED] exporter batcher - byte size based batching