--- a/pdata/plog/generated_scopelogs.go
+++ b/pdata/plog/generated_scopelogs.go
@@ -46,6 +46,10 @@ func (ms ScopeLogs) MoveTo(dest ScopeLogs) {
 	*ms.orig = otlplogs.ScopeLogs{}
 }
 
+func (ms ScopeLogs) Size() int {
+	return ms.orig.Size()
+}
+
 // Scope returns the scope associated with this ScopeLogs.
 func (ms ScopeLogs) Scope() pcommon.InstrumentationScope {
 	return pcommon.InstrumentationScope(internal.NewInstrumentationScope(&ms.orig.Scope, ms.state))