--- a/pdata/pmetric/generated_exemplar.go
+++ b/pdata/pmetric/generated_exemplar.go
@@ -50,6 +50,10 @@ func (ms Exemplar) MoveTo(dest Exemplar) {
 	*ms.orig = otlpmetrics.Exemplar{}
 }
 
+func (ms Exemplar) Size() int {
+	return ms.orig.Size()
+}
+
 // Timestamp returns the timestamp associated with this Exemplar.
 func (ms Exemplar) Timestamp() pcommon.Timestamp {
 	return pcommon.Timestamp(ms.orig.TimeUnixNano)