--- a/pdata/pmetric/generated_resourcemetrics.go
+++ b/pdata/pmetric/generated_resourcemetrics.go
@@ -46,6 +46,10 @@ func (ms ResourceMetrics) MoveTo(dest ResourceMetrics) {
 	*ms.orig = otlpmetrics.ResourceMetrics{}
 }
 
+func (ms ResourceMetrics) Size() int {
+	return ms.orig.Size()
+}
+
 // Resource returns the resource associated with this ResourceMetrics.
 func (ms ResourceMetrics) Resource() pcommon.Resource {
 	return pcommon.Resource(internal.NewResource(&ms.orig.Resource, ms.state))