--- a/pdata/pprofile/generated_resourceprofiles.go
+++ b/pdata/pprofile/generated_resourceprofiles.go
@@ -46,6 +46,10 @@ func (ms ResourceProfiles) MoveTo(dest ResourceProfiles) {
 	*ms.orig = otlpprofiles.ResourceProfiles{}
 }
 
+func (ms ResourceProfiles) Size() int {
+	return ms.orig.Size()
+}
+
 // Resource returns the resource associated with this ResourceProfiles.
 func (ms ResourceProfiles) Resource() pcommon.Resource {
 	return pcommon.Resource(internal.NewResource(&ms.orig.Resource, ms.state))