--- a/pdata/ptrace/generated_spanlink.go
+++ b/pdata/ptrace/generated_spanlink.go
@@ -49,6 +49,10 @@ func (ms SpanLink) MoveTo(dest SpanLink) {
 	*ms.orig = otlptrace.Span_Link{}
 }
 
+func (ms SpanLink) Size() int {
+	return ms.orig.Size()
+}
+
 // TraceID returns the traceid associated with this SpanLink.
 func (ms SpanLink) TraceID() pcommon.TraceID {
 	return pcommon.TraceID(ms.orig.TraceId)