… exporter/exporterhelper/internal/retry_sender_test.go
… exporter/exporterhelper/logs_batch_test.go
… exporter/exporterhelper/xexporterhelper/profiles_batch_test.go
… exporter/exporterqueue/persistent_queue_test.go
… exporter/go.sum
… exporter/internal/queue/fake_request_test.go