@sfc-gh-sili 2025-01-14
Feedback from Bogdan: